Added ShoppPages implementation to replace StorefrontPages
The ShoppPages approach instantiates all storefront pages for every
request so that page links are available. With only a handful 5-6
pages, this is not a major memory hit, but with the running page models
it is possible to translate the page titles and slugs through WPML and
consolidate the interfaces to decouple the components.