[#2179] Initial work backporting the registration form fixes to 1.2.x.
Rather than using checkout.js on the registration page (we'd then need to resolve other issues relating to dependency resolution as it is enqueued late in the request) address.js from 1.3 has been added in with a single dependency on shopp.js.

Work still to be done: testing the registration process completely.