#2226 refactored Adam's patch re WP user account creation.
The key change is that if no email/an invalid email addr is provided and/or no password the user account creation will fail (whether WP integration is enabled or not).

Still todo: if WP integration enabled, do we abort Shopp account creation if WP account creation fails for any reason?