Fixed #2110 Removed the unset joins call in facet query
The unset joins call is supposed to remove the joins in the query to reduce the query complexity (speeding it up) since faceted menus have a covering taxonomy query to isolate the set of matching products. The side effect of removing the joins is that it breaks the taxonomy-based sort order methods custom and recommended which expect a object with a ProductTaxonomy ancestry to have the joins available for sorting against.