Fixed #944 and fixed #1953 so that taxes are (re)calculated after discounts are applied
Discounts are applied pre-tax such that the tax amount is calculated on the discounted price.

The changes in this commit refactor the ShoppTax::rates() method to process the rates by reference, maintaining the original data and updating it so that it may be properly used by reference in the OrderAmountItemTax class for properly recalculating the tax amounts after discounts are applied.

From http://www.hmrc.gov.uk/vat/managing/charging/discounts-etc.htm

> **Discounts on goods and services**
> If any of your goods or services are discounted, you charge VAT on the discounted price rather than the full price.
> If you make an offer to a customer such as 'we will pay your VAT', VAT is actually payable to HM Revenue & Customs (HMRC) on the amount the customer would have paid on the discounted price, not the amount they have paid at the full price.
> **Free gifts**
> If you give away products, you must usually account for, and pay VAT on, the full value or cost of the gift, although there are some exceptions.