Updated script version values for better cache control; Replaced SWFUpload with a secured version
The SWFUpload JS files have been removed from Shopp (to use those that ship with WordPress), but the swfupload.swf is still be distributed with Shopp since we use a modified version that incorporates the latest security fixes and a bug fix to allow `button_text_style` to work.

The htmlText value of the button was changed to include XSS protection, and wrapped in a standard `<span class="button-text">`. The issue is that AS3 cannot handle hyphenated CSS class names. Changing the class to `buttonText` allows it to work with the `button_text_style` parameter.