Tweak to comparison string (needed because of addons created in an earlier test)
Not happy totally happy with this, it seems like a brittle approach and for instance test_product_variation_tags fails if I run ProductAPITests by themselves.

The affected parts of the comparison string are the IDs... perhaps we need helpers to reliably determine what these are, regardless of changes during the setup phase of other test classes?

I do feel we should be able to run any combination of these tests - and they should standalone and succeed.