For #2154 If the billing address is empty, copy the shipping address when it is updated during cart estimates
This assumes that if the address, xaddress and city properties are empty in the BillingAddress, it is "empty". If the address, xaddress or city are filled in, ShippingAddress updates during cart estimates are ignored.

This also allows for multiple cart estimate updates so that if the post code or country are changed in the cart, the billing address will keep shifting up until an actual address (street and city) are provided.