Refactored the checkout workflow for PayPal Standard
After submitting the order for PayPal payment, the normal confirm order process runs. Instead of again needing to click the "Checkout with PayPal" button, a redirect page is used to auto-submit the POST form to submit the order to PayPal from the client browser.

If JavaScript is disabled, graceful degradation provides another "Checkout with PayPal" submit button to submit the order for payment.

This process brings offsite payment processing much closer to the same process as onsite payments.