Fixed #2517 Refactored all static callbacks to use array notation; Corrected ProductCollection::page writes() to be declared static
String notation for static callbacks was not added until PHP 5.2.3 (according to the PHP Docs). Additionally, string notation callbacks for static methods will fail if the referenced method is not explicitly declared as static (static keyword must appear in the function declaration).

As a point of convention, we will use array notation for all object/class callbacks.