#2630 Included migration script to fix inclusive tax prices to re-include the tax amounts
Shopp 1.1 introduced the "base price" system where prices entered into the product editor would be stripped of inclusive taxes to store the tax excluded price in the database record. This was, at the time, thought to provide more flexibility for moving to different tax rates, etc.

A more complete understanding of inclusive taxes shows that with inclusive tax systems, the tax is taken as a ratio (proportion) of the price. The price is still the price, and yes, tax rate changes change the ratio of net earnings to tax costs. This is expected in Europe and is very natural there where merchants pad prices for their country tax.

This change set uses the base of operations to determine what tax rate applied in the product editor to add those taxes back in to the prices. The inclusive tax adjustments also support tax conditions for category and tag taxonomies to ensure that reduced rate taxes are applied to products where appropriate. NOTE: this only works for conditions matching ANY condition, not ALL conditions.

Additionally, a temporary fix is in place to reset everyone's DB version to the correct numbering to ensure the 1.3 upgrade scripts run. This adjustment should be removed from a future public release.