Fixed #2630 Implemented new inclusive tax behaviors
When inclusive taxes are enabled, the tax amount is no longer included in the Totals register calculations as the value of determined tax is simply a calculation derived from the item prices (and/or shipping). Since these figures are already part of the price, the tax total should not be added "again" by the totals register calculations.

A lot of Theme API calls were refactored to streamline and simplify the maintenance process of the new tax behaviors.