Fixed product.image and product.coverimage to properly handle data loading requests
When loading just a cover image, it is important that a load_data request for 'images' will override the cover images requests and load all the images for the product so that the Theme API will respond as the developer expects it to.

The 'coverimage' request is a special case because it is designed to only retrieve a single entry that goes into the same 'images' property as when loading multiple images. That would make the other images calls believe that all images available were already loaded.

We now detect keep track of what data has already been requested to prevent from re-querying when unnecessary.