Fixed #2737 Manually fire ShoppOrder->success() if the order is already captured from IPN
When IPN messages are processed before PDT, the success() method has no way of running. Additionally, reworked the pdt() method remove it's reliance on ShoppOrder->inprogress. It uses the authenticated PDT message to get transaction details and load the proper order to work on. It then needs to replace the ShoppOrder->inprogress property so that when ShoppOrder->success() runs the correct order is marked as completed.