Fixed #2846 Moved  checkout request redirect to only occur with shopping cart requests
In particular, we must make sure this only runs after $_REQUEST['cart'] is present ( ! empty() ) so that we are sure it only occurs if a cart request is present and detected. In 1.2, this occurred in the Cart->request() handler which fired on the shopp_cart_request action hook.

For this fix, we're moving it to after the shopp_cart_request hook so that it is in a similar location to where it processed before, but living in different code (the super-controller in this case).