Delete dimension entries from shopp_meta on product deletion
In table shopp_meta, dimensions are stored with a relation to table shopp_price on shopp_meta.parent = shopp_price.id. 

Right now these entries in meta are left behind after deleting a product. Because of the mentioned relation the removal of the entries needs to be done before removal of shopp_price entries.