Fixed #1995 Added discount support for PayPal Standard subscriptions
Discounts on subscriptions are not straightforward since it is unclear whether to add the discount to the ongoing subscription price, or the introductory trial offer only. By default, Shopp will apply the discount to the ongoing subscription price. Filtering the shopp_paypalstandard_discount_trials to return boolean true will apply the discounts to the trial price instead.