Fixed #3063 State input behavior fixed for order admin address edit
The address editor was not correctly setup for managing different editor types (billing vs shipping) and the behaviors were not able to find the editor elements properly.

In addition the editor UI is not directly appended to the DOM until have the state menu/input behavior is initialized. This meant DOM traversal has to be done relative to the country menu element not against the global Document DOM.