Fixed #3101 Bind storefront scripts to shopp.js
In addition, added jQuery dependency to address.js and ensured that address.js will not break if no regions variable is provided.

Theoretically this shouldn't be possible anyways since the global scripts have been bound to the shopp.js which is a dependency of address.js. The same goes for global script settings for checkout.js which also has a shopp.js dependency.

Before, the storefront scripts were bound to catalog.js. Since catalog.js was not a dependency of these other scripts, if you deregistered catalog.js, the global scripts would not be set and the entire JS environment would crash on undefined references. This is why the storefront global JS was moved to be bound on the more universally dependent shopp.js.