Added ShoppServices must-use-plugin (mu-plugin)
This plugin replaces the need for wp-load.php scanning though we lose the performance gains of using SHORTINIT. At the same time, the purpose of the plugin is to load early enough to block loading other plugins that may bog down service requests while still loading the core of Shopp necessary to support the service. This makes service requests for our various resource server scripts (images, scripts, styles) as efficient and reliable as possible.

Props to WP Migrate Pro for developing the technique and clifgriffin for the recommendation.