Fixed #3126 Payment field processing must come after billaddress
The billaddress() method includes a call to $this->Register->billaddress() which is responsible for copying the form data en masse to the billing object. If payment processing happens ahead of Register->billaddress() the form data will replace the already processed payment data.

This results in symptoms as seen in issue #3126.