Closed #3144 Generate new sessions for orphan session cookies
If a session cookie exists, but no session record exists, a new session needs assigned to ensure the old session cookies are replaced. This allows a merchant to wipe the session table and recreate sessions for all users. It also reduces and prevents an edge case where an old cookie might get used for an existing session.