Fixed #3203 Ensure the order lock is destroyed during shutdown
If execution of PHP suddenly halts (PHP Fatal error, for instance) it will leave the lock in place, blocking other processes from being able to manipulate the order in the session. To prevent this, a handler is added to delete the transient during the WP shutdown action, which fires whenever the PHP process is ending. This ensures locks are always cleared after the PHP processing ends.