Fixed #3197 Update shipped property on load
A ShoppCartItem->shipped property must be able to be set to true or false during load in order to deal with the variant change context in the cart. Prior the logic was only setting it true when the origin product's type is detected as 'Shipped'.

The cart's change() method, however, uses ShoppCartItem->load() to repopulate the item flags fresh from the origin product. Without being able to override from the product, the shipped property would stay locked to true, never allowing the Shiprates to recalculate.