Fixed #3201 Check the truthishness of the show product option
The widget saves boolean values in strings 'on' or 'off'. The logic to check if the setting was on was insufficiently using the boolean conversion of a string (true if not empty) to detect if it was set in the ShoppCategoryDropdownWalker.

Shopp::str_true() correctly converts the string values of 'on' or 'off' to boolean values.