Lookup gateway using Shopp 1.3.x naming scheme on Orders admin page
The gateway field of the Orders admin is blank after upgrading to Shopp
1.3.x (currently 1.3.9dev) for orders created before upgrading.  The
reason is that the class names of the gateways have changed to include a
"Shopp" prefix.

If a gateway's name is not found as stored with the Order, try to look
it up again with the "Shopp" prefix.

https://github.com/ingenesis/shopp/issues/3253