Restore "Charge Order" button for orders placed before Shopp 1.3.x
The "Charge Order" button is missing from the Order detail admin page in
Shopp 1.3.x for orders placed on 1.2.5.  The reason is that the class
names of the gateways have changed to include a "Shopp" prefix, so the
Gateway lookup fails.

If a gateway's name is not found as stored with the Purchase record, try
to look it up again with the "Shopp" prefix.

https://github.com/ingenesis/shopp/issues/3253