Update tagcloud widget
Update tagcloud widget with exclude option.

The tags are often used for other things than just terms that relate to the product. For example, when two tax rates are in use, you could tag a product 'high' or 'low' to reflect the proper tax rate for the product. When using the tagcloud widget you don't want the phrases 'high' and/or 'low' to show up. Adding an exclude option in the widget makes it easier to deal with this.