Fix weak typing evaluation in copydata()
This is a pretty major bug as any copydata() use that passes in an empty array ends up with all of the default ignore properties.

This particuarly caused a bug with the Login controller wherein the loaded ShoppCustomer object did not have the customer’s ID set, causing the My Account screens to malfuction.